ai_dev_team/requirements.txt
crewai==0.28.8
langgraph==0.0.44
python-dotenv
requests==2.31.0
streamlit==1.38.0
langgraph
dev_agents
pytest==8.3.2
beautifulsoup4==4.12.3
lxml==5.3.0
structlog==24.1.0